Move tools methods into service layer
Add a service layer for the API services for the /tools page.
